requirements.txt
# ✅ DEPENDÊNCIAS CONSOLIDADAS - FORENSENET v2.3

# Banco de Dados
sqlalchemy>=2.0
psycopg2-binary>=2.9

# API Moderna
fastapi>=0.100
uvicorn>=0.23

# Filas e Message Broker
celery>=5.3
redis>=5.0

# IA Local & NLP
transformers>=4.35
torch>=2.1
sentence-transformers>=2.2
spacy>=3.7
nltk>=3.8
scikit-learn>=1.3

# Scraping Resiliente
instagrapi>=2.1
playwright>=1.40
playwright-stealth>=1.0
beautifulsoup4>=4.12
requests>=2.31

# Visualização & Dados
pandas>=2.0
numpy>=1.25
matplotlib>=3.8
seaborn>=0.12
wordcloud>=1.9

# Observabilidade & Utilitários
python-dotenv>=1.0
structlog>=23.2
prometheus-client>=0.19
tqdm>=4.66
httpx>=0.24
gunicorn>=21.2
boto3>=1.28
